requirements.txt
pip>=22.2.2
torch
torchvision
pandas
pytorch-lightning
tensorboard
tabulate
tqdm
pillow
notebook
jupyterlab
matplotlib
seaborn
ipywidgets
boto3
pyprojroot
scikit-learn
opencv-python
wandb